requirements.txt
Django==1.7.2
MySQL-python==1.2.5
Pillow==2.7.0
South==1.0.2
beautifulsoup4==4.3.2
django-copyright==1.0.0
django-happenings==0.2.1
django-tinymce==1.5.3
oauthlib==0.7.2
pytz==2014.10
requests==2.5.1
requests-oauthlib==0.4.2
wsgiref==0.1.2
